Move model index based completer to local module
